requirements.txt
# Django Core
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0

# Environment and Configuration
python-dotenv>=1.0.0

# Image Processing (for photo verification)
Pillow>=10.0.0

# Security and Encryption
cryptography>=41.0.0

# Photo Verification and AI
face-recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR for ID document text extraction
pytesseract>=0.3.10

# Production Server
gunicorn>=21.0.0
whitenoise>=6.0.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Firebase (already in use)
firebase-admin>=6.0.0

# HTTP requests
requests>=2.31.0